+++ app/Http/Controllers/SaleController.php
@@ -43,7 +43,7 @@ class SaleController extends Controller
         $days_ago = Carbon::now()->subDays(5);
 
         // Obtener las ventas registradas en los últimos x días
-        $sales = Sale::whereDate('created_at', '>=', $days_ago)->latest()->get();
+        $sales = Sale::where('store_id', auth()->user()->store_id)->whereDate('created_at', '>=', $days_ago)->latest()->get();
 
         // Agrupar las ventas por fecha con el nuevo formato de fecha y calcular el total de productos vendidos y el total de ventas para cada fecha
         $groupedSales = $sales->groupBy(function ($sale) {
@@ -126,7 +126,7 @@ class SaleController extends Controller
         $date = Carbon::parse($created_at)->toDateString();
 
         // Obtener las ventas registradas en la fecha recibida
-        $sales = Sale::with(['product:id,name', 'globalProductStore.globalProduct'])->whereDate('created_at', $date)->get();
+        $sales = Sale::where('store_id', auth()->user()->store_id)->with(['product:id,name', 'globalProductStore.globalProduct'])->whereDate('created_at', $date)->get();
 
         // return $sales;
 
@@ -216,7 +216,7 @@ class SaleController extends Controller
         $saleDate = $sale->created_at->toDateString();
 
         // Eliminar todos los registros que tengan la misma fecha de creación
-        Sale::whereDate('created_at', $saleDate)->delete();
+        Sale::where('store_id', auth()->user()->store_id)->whereDate('created_at', $saleDate)->delete();
 
         // Eliminar el registro de venta enviado como referencia
         $sale->delete();
@@ -230,7 +230,7 @@ class SaleController extends Controller
         $endDate = Carbon::parse($queryDate[1])->endOfDay();
 
         // Obtener las ventas registradas en el rango de fechas requerido por el filtro
-        $sales = Sale::whereDate('created_at', '>=', $startDate)->whereDate('created_at', '<=', $endDate)->latest()->get();
+        $sales = Sale::where('store_id', auth()->user()->store_id)->whereDate('created_at', '>=', $startDate)->whereDate('created_at', '<=', $endDate)->latest()->get();
 
         // Agrupar las ventas por fecha con el nuevo formato de fecha y calcular el total de productos vendidos y el total de ventas para cada fecha
         $groupedSales = $sales->groupBy(function ($sale) {
@@ -254,8 +254,8 @@ class SaleController extends Controller
 
     public function getItemsByPage($currentPage)
     {
-        $offset = 5 + $currentPage * 5; //multiplica por 2 para traer de 2 dias en 2 dias. suma 5 dias porque son los que ya se cargaron
-        $skip_days = $currentPage * 5; //multiplica por 2 para traer de 2 dias en 2 dias
+        $offset = 5 + $currentPage * 5; //multiplica por 5 para traer de 5 dias en 5 dias. suma 5 dias porque son los que ya se cargaron
+        $skip_days = $currentPage * 5; //multiplica por 5 para traer de 5 dias en 5 dias
 
          // Calcular la fecha hace x días para recuperar las ventas de x dias atras hasta la fecha de hoy
          $days_ago = Carbon::now()->subDays($offset);
@@ -263,7 +263,7 @@ class SaleController extends Controller
          $days_befor = Carbon::now()->subDays($skip_days);
 
          // Obtener las ventas registradas en los últimos 7 días
-         $sales = Sale::whereDate('created_at', '>=', $days_ago)->whereDate('created_at', '<=', $days_befor)->latest()->get();
+         $sales = Sale::where('store_id', auth()->user()->store_id)->whereDate('created_at', '>=', $days_ago)->whereDate('created_at', '<=', $days_befor)->latest()->get();
  
          // Agrupar las ventas por fecha con el nuevo formato de fecha y calcular el total de productos vendidos y el total de ventas para cada fecha
          $groupedSales = $sales->groupBy(function ($sale) {
